demical task ratings

+++ src/main/java/com/mas/dashboard/entity/TaskRating.java
@@ -24,7 +24,7 @@ public class TaskRating {
 
   private Long studentId;
 
-  private float rating;
+  private Double rating;
 
   private Boolean deleted;
